（1）创建群之后，默认群主在群里，该群聊自动加入聊天队列 （2）退群时判断退群人是否是群主，如果是需要转让群主；如果退群前群里只有一个人，退出成功后在聊天队列中删除该群

+++ server/src/main/java/com/itkey/chatroom/service/impl/RoomServiceImpl.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 import com.itkey.chatroom.VO.ResultVO;
+import com.itkey.chatroom.repository.MessageRepository;
 import com.itkey.chatroom.utils.ResultVOUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -32,6 +33,8 @@ public class RoomServiceImpl {
     @Autowired
     private UserInRoomRepository userInRoomRepository;
 
+    @Autowired
+    private MessageRepository messageRepository;
     /**
      * 创建群
      *
@@ -56,6 +59,12 @@ public class RoomServiceImpl {
             return ResultVOUtil.error(-3, "创建群，设置个头像吧！");
         }
         room.setOwner(user);
+        //创建群默认把自己加入到群中去
+        UserInRoom userInRoom = new UserInRoom();
+        userInRoom.setUser(user);
+        userInRoom.setRoom(room);
+        userInRoomRepository.save(userInRoom);
+
         return ResultVOUtil.success(roomRepository.save(room));
     }
 
@@ -94,6 +103,33 @@ public class RoomServiceImpl {
         } else {
             return ResultVOUtil.error(-4, "数据错误！");
         }
+    }
+
+
+    /**
+     * 更改群主
+     *
+     * @param userId 用户信息
+     * @param roomId 群id
+     * @return
+     */
+    public ResultVO transferRoom(Long userId, Long roomId) {
+        Optional<User> userOptional = userRepository.findById(userId);
+        Optional<Room> roomOptional = roomRepository.findById(roomId);
+
+        if (!userOptional.isPresent()) {
+            return ResultVOUtil.error(-1, "用户不存在！");
+        }
+        if (!roomOptional.isPresent()) {
+            return ResultVOUtil.error(-2, "错误的群id！");
+        }
+        //判断用户是否已经加过群了
+        User user = userOptional.get();
+        Room room = roomOptional.get();
+        //todo: 实际使用时必须增加权限校验
+        room.setOwner(user);
+        Room result = roomRepository.save(room);
+        return ResultVOUtil.success(result);
     }
 
     /**
@@ -193,7 +229,76 @@ public class RoomServiceImpl {
         if (!roomOptional.isPresent()) {
             return ResultVOUtil.error(-4, "[" + roomId + "]房间不存在！");
         }
+        Room room = roomOptional.get();
+
+        //判断当前群是否只有一个人，如果只有一个人则提示退出群会自己删除群和聊天记录
+        int userCount = userInRoomRepository.countByRoom(room);
+        if(userCount<=1){
+            return ResultVOUtil.error(-5, "您是群里的唯一成员，您退出群以后，群将不存在，所有聊天记录将被删除！");
+        }
+        //判断退群的人是否是群主，如果是群主则提示需要先转让群
+        if(room.getOwner().getId().equals(userOptional.get().getId())){
+            return ResultVOUtil.error(-6, "您是该群的群主，请先转让该群方可退出群聊。");
+        }
+
         userInRoomRepository.deleteByUserAndRoom(userOptional.get(),roomOptional.get());
+        return ResultVOUtil.success();
+    }
+    /**
+     * 群中只有一个人的情况，单独处理
+     * @param roomId
+     * @param userId
+     * @return
+     */
+    @Transactional
+    public ResultVO exitRoomOnlyOne(Long roomId, Long userId) {
+        //TODO: 和上面代码重复比较多，可优化或者合并
+        if(roomId==null){
+            return ResultVOUtil.error(-1, "参数roomId不能为空!");
+        }
+        if(userId==null){
+            return ResultVOUtil.error(-2, "参数userId不能为空!");
+        }
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (!userOptional.isPresent()) {
+            return ResultVOUtil.error(-3, "[" + userId + "]用户不存在！");
+        }
+        Optional<Room> roomOptional = roomRepository.findById(roomId);
+        if (!roomOptional.isPresent()) {
+            return ResultVOUtil.error(-4, "[" + roomId + "]房间不存在！");
+        }
+        Room room = roomOptional.get();
+
+        //判断当前群是否只有一个人，如果只有一个人则提示退出群会自己删除群和聊天记录
+        int userCount = userInRoomRepository.countByRoom(room);
+        if(userCount!=1){
+            return ResultVOUtil.error(-5, "用户人数不为1，执行失败！");
+        }
+        //判断退群的人是否是群主，如果是群主则提示需要先转让群
+        userInRoomRepository.deleteByUserAndRoom(userOptional.get(),room);
+        //删除群的所有聊天记录
+        deleteRoomAndRoomMessage(roomId);
+        return ResultVOUtil.success();
+    }
+
+    /**
+     * 删除群和群记录
+     * @param roomId 群ID
+     * @return
+     */
+    @Transactional
+    public ResultVO deleteRoomAndRoomMessage(Long roomId){
+        Optional<Room> roomOptional = roomRepository.findById(roomId);
+        if (!roomOptional.isPresent()) {
+            return ResultVOUtil.error(-4, "[" + roomId + "]群不存在！");
+        }
+        Room room = roomOptional.get();
+        //删除聊天记录
+
+        messageRepository.deleteByRoomId(roomId);
+
+        //删除房间信息
+        roomRepository.delete(room);
         return ResultVOUtil.success();
     }
 }
